execute regardless of working dir

+++ run.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+cd "$SCRIPT_DIR"
+
 start_time=$SECONDS
 
 echo -en "\n"
@@ -16,8 +19,15 @@ echo "     RUNNING     "
 echo "################"
 echo -en "\n"
 
-setterm -linewrap off
-stdbuf -oL nice -n 19 ./build/net 2>&1 | tee out
+
+[ -w out ] && stdbuf -oL nice -n 19 ./build/net 2>&1 | tee out || ./build/net
+
+if { [ -e out ] && [ -w out ]; } || { touch out && [ -w out ]; }; then
+    stdbuf -oL nice -n 19 ./build/net 2>&1 | tee out
+else
+    ./build/net
+fi
+
 
 elapsed_time=$((SECONDS - start_time))
 echo -en "Elapsed time: $elapsed_time seconds\n" | tee -a out
